requirements.txt
streamlit
pandas
matplotlib
fpdf2
openpyxl
